- move the custom collections to an upper level package
