+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/collections/PhoenicisTransformationList.java
@@ -1,4 +1,4 @@
-package org.phoenicis.javafx.views.common.lists;
+package org.phoenicis.javafx.collections;
 
 import javafx.collections.ListChangeListener;
 import javafx.collections.ObservableList;
