+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/common/widgets/lists/compact/CompactListWidget.java
@@ -5,7 +5,7 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.scene.control.ListView;
 import javafx.scene.input.MouseEvent;
-import org.phoenicis.javafx.views.common.lists.MappedList;
+import org.phoenicis.javafx.collections.MappedList;
 import org.phoenicis.javafx.views.common.widgets.lists.ListElementListCell;
 import org.phoenicis.javafx.views.common.widgets.lists.ListWidget;
