+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/mainwindow/apps/ApplicationPanel.java
@@ -28,7 +28,7 @@ import javafx.scene.layout.*;
 import javafx.scene.web.WebView;
 import org.phoenicis.javafx.settings.JavaFxSettingsManager;
 import org.phoenicis.javafx.views.common.ErrorMessage;
-import org.phoenicis.javafx.views.common.lists.PhoenicisFilteredList;
+import org.phoenicis.javafx.collections.PhoenicisFilteredList;
 import org.phoenicis.javafx.views.common.ThemeManager;
 import org.phoenicis.javafx.views.common.widgets.lists.DetailsView;
 import org.phoenicis.repository.dto.ApplicationDTO;
